Page edits

+++ src/pages/ProductPage.js
@@ -2,10 +2,12 @@ import React, { useState, useEffect } from "react";
 import { useParams, Link } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
 import Navbar from "../components/Navbar";
+import Footer from "../components/Footer";
 
 const ProductPage = () => {
   const [product, setProduct] = useState({});
   const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [isAuthorized, setIsAuthorized] = useState(false);
   const [quantity, setQuantity] = useState(1);
   const { id } = useParams();
 
@@ -13,9 +15,18 @@ const ProductPage = () => {
     const token = localStorage.getItem("token");
     if (token) {
       setIsLoggedIn(true);
+      try {
+        const decodedToken = jwtDecode(token);
+        const userName = decodedToken.name;
+        setIsAuthorized(userName === "admin");
+      } catch (error) {
+        console.error("Error decoding token:", error);
+        setIsLoggedIn(false);
+        setIsAuthorized(false);
+      }
     }
 
-    fetch(`https://mw-project-be.vercel.app/product/${id}`)
+    fetch(`http://localhost:8000/product/${id}`)
       .then((response) => response.json())
       .then((data) => {
         console.log(data);
@@ -30,22 +41,24 @@ const ProductPage = () => {
 
   const handleDelete = async () => {
     try {
-      const response = await fetch(
-        `https://mw-project-be.vercel.app/product/${id}`,
-        {
+      const token = localStorage.getItem("token");
+      if (token) {
+        const response = await fetch(`http://localhost:8000/product/${id}`, {
           method: "DELETE",
           headers: {
             Authorization: token,
           },
+        });
+        if (response.ok) {
+          const data = await response.json();
+          console.log(data.message);
+          alert("Successfully deleted");
+          window.location.href = "/";
+        } else {
+          console.error("Error deleting product:", response.statusText);
         }
-      );
-      if (response.ok) {
-        const data = await response.json();
-        console.log(data.message);
-        alert("Successfully deleted");
-        window.location.href = "/";
       } else {
-        console.error("Error deleting product:", response.statusText);
+        console.error("No token found in local storage");
       }
     } catch (error) {
       console.error("Error deleting product:", error);
@@ -53,6 +66,7 @@ const ProductPage = () => {
   };
 
   const handleAddToCart = () => {
+
     let savedCart = localStorage.getItem("cart");
     if (savedCart) {
       let parsedCart = JSON.parse(savedCart);
@@ -77,74 +91,126 @@ const ProductPage = () => {
         ? product.title.slice(0, 25) + "..."
         : product.title;
 
-    alert(`${pr_title} (${product.quantity}) was added to cart!`);
-    //temporary alert - add an cart icon that has the number of items (in the cart) displayed for user if there is at least one item
+    if(!isLoggedIn){
+      window.location.href = "/login";
+      alert("Please login to access your added items!")
+    }
+    else {
+      alert(`${pr_title} was added to the cart!`);
+    }
   };
 
-  const token = localStorage.getItem("token");
-  const decodedToken = jwtDecode(token);
-  const userName = decodedToken.name;
-
-  const isAuthorized = isLoggedIn && userName === "admin";
-
   return (
-    <div className="book-container-book">
+    <div className="product-container">
       <Navbar />
-      <img src={product.image} width={250} alt="product_image" />
-
-      <h1 className="book-title">{product.title}</h1>
-      <h5 className="book-author">CAD$ {product.price}</h5>
-
-      <div className="book-details">
-        <p>Category: {product.category}</p>
-        <p>Rating: {product.rating}</p>
-        <p>Quantity: {product.quantity}</p>
-        <p>Description: {product.description}</p>
-      </div>
-
-      <div className="book-btn-container">
-        {isAuthorized && (
-          <div>
-            <Link
-              className="book-btn edit-book-btn"
-              to={`/product/edit/${product._id}`}
-            >
-              Edit Prodcut
-            </Link>
-            <button
-              className="book-btn delete-book-btn"
-              data-id={product._id}
-              onClick={handleDelete}
-            >
-              Delete Product
-            </button>
+      <div
+        className="product-details"
+        style={{
+          display: "flex",
+          alignItems: "center",
+          gap: "20px",
+          padding: "20px",
+          backgroundColor: "white",
+          boxShadow: "0 2px 4px rgba(0, 0, 0, 0.1)",
+          borderRadius: "5px",
+        }}
+      >
+        <img src={product.image} width={250} alt="product_image" />
+        <div>
+          <h1 className="product-title">{product.title}</h1>
+          <h5 className="product-price">CAD$ {product.price}</h5>
+          <div className="product-details">
+            <p>Category: {product.category}</p>
+            <p>Rating: {product.rating}</p>
+            <p>Quantity: {product.quantity}</p>
+            <p>Description: {product.description}</p>
           </div>
-        )}
-        <div
-          style={{
-            display: "flex",
-            width: 150,
-            justifyContent: "space-between",
-          }}
-        >
-          <button
-            onClick={() => {
-              if (quantity > 1) setQuantity(quantity - 1);
-            }}
-          >
-            -
-          </button>
-          <div>{quantity}</div>
-          <button
-            onClick={() => {
-              setQuantity(quantity + 1);
-            }}
-          >
-            +
-          </button>
-          <button onClick={handleAddToCart}>Add to Cart</button>
+          <div className="product-actions">
+            {isAuthorized && (
+              <div>
+                <Link
+                  className="product-btn edit-product-btn"
+                  to={`/product/edit/${product._id}`}
+                  style={{
+                    backgroundColor: "#39b575",
+                    color: "white",
+                    textDecoration: "none",
+                    padding: "5px 10px",
+                    borderRadius: "5px",
+                  }}
+                >
+                  Edit Product
+                </Link>
+                <button
+                  className="product-btn delete-product-btn"
+                  data-id={product._id}
+                  onClick={handleDelete}
+                  style={{
+                    backgroundColor: "#39b575",
+                    color: "white",
+                    border: "none",
+                    padding: "5px 10px",
+                    borderRadius: "5px",
+                  }}
+                >
+                  Delete Product
+                </button>
+              </div>
+            )}
+            <div
+              style={{
+                display: "flex",
+                width: 150,
+                justifyContent: "space-between",
+                alignItems: "center",
+              }}
+            >
+              <button
+                onClick={() => {
+                  if (quantity > 1) setQuantity(quantity - 1);
+                }}
+                style={{
+                  backgroundColor: "#39b575",
+                  color: "white",
+                  border: "none",
+                  padding: "5px 10px",
+                  borderRadius: "5px",
+                }}
+              >
+                -
+              </button>
+              <div>{quantity}</div>
+              <button
+                onClick={() => {
+                  setQuantity(quantity + 1);
+                }}
+                style={{
+                  backgroundColor: "#39b575",
+                  color: "white",
+                  border: "none",
+                  padding: "5px 10px",
+                  borderRadius: "5px",
+                }}
+              >
+                +
+              </button>
+              <button
+                onClick={handleAddToCart}
+                style={{
+                  backgroundColor: "#39b575",
+                  color: "white",
+                  border: "none",
+                  padding: "5px 10px",
+                  borderRadius: "5px",
+                }}
+              >
+                Add to Cart
+              </button>
+            </div>
+          </div>
         </div>
       </div>
+      <Footer pos={"absolute"}/>
     </div>
   );
 };
